+++ src/dcos_e2e_cli/dcos_aws/commands/_common.py
@@ -12,6 +12,7 @@ from dcos_e2e.backends import AWS
 from dcos_e2e.cluster import Cluster
 from dcos_e2e.distributions import Distribution
 from dcos_e2e.node import Node, Role
+from dcos_e2e_cli._vendor.dcos_launch import config, get_launcher
 from dcos_e2e_cli.common.base_classes import ClusterRepresentation
 
 CLUSTER_ID_TAG_KEY = 'dcos_e2e.cluster_id'
@@ -55,7 +56,8 @@ def existing_cluster_ids(aws_region: str) -> Set[str]:
     """
     ec2 = boto3.resource('ec2', region_name=aws_region)
     ec2_filter = {'Name': 'tag:' + CLUSTER_ID_TAG_KEY, 'Values': ['*']}
-    ec2_instances = ec2.instances.filter(Filters=[ec2_filter])
+    state_filter = {'Name': 'instance-state-name', 'Values': ['running']}
+    ec2_instances = ec2.instances.filter(Filters=[ec2_filter, state_filter])
 
     cluster_ids = set()  # type: Set[str]
     for instance in ec2_instances:
@@ -97,7 +99,8 @@ class ClusterInstances(ClusterRepresentation):
             'Name': 'tag:' + NODE_TYPE_TAG_KEY,
             'Values': [node_types[role]],
         }
-        filters = [cluster_id_tag_filter, node_role_filter]
+        state_filter = {'Name': 'instance-state-name', 'Values': ['running']}
+        filters = [cluster_id_tag_filter, node_role_filter, state_filter]
         ec2_instances = set(ec2.instances.filter(Filters=filters))
         return ec2_instances
 
@@ -220,7 +223,53 @@ class ClusterInstances(ClusterRepresentation):
     def destroy(self) -> None:
         """
         Destroy this cluster.
-
-        This is not yet implemented, see:
-        https://jira.mesosphere.com/browse/DCOS_OSS-5042
         """
+        backend = AWS()
+        deployment_name = self._cluster_id
+        masters = len(self.masters)
+        agents = len(self.agents)
+        public_agents = len(self.public_agents)
+
+        # We need this to be set but not necessarily correct.
+        aws_instance_type = backend.aws_instance_type
+
+        launch_config = {
+            'admin_location': backend.admin_location,
+            'aws_region': self._aws_region,
+            'deployment_name': deployment_name,
+            'installer_url': 'https://example.com',
+            'instance_type': aws_instance_type,
+            'launch_config_version': 1,
+            'num_masters': masters,
+            'num_private_agents': agents,
+            'num_public_agents': public_agents,
+            'platform': 'aws',
+            'provider': 'onprem',
+        }
+
+        launch_config['dcos_config'] = backend.base_config
+        validated_launch_config = config.get_validated_config(
+            user_config=launch_config,
+            config_dir=str(self._workspace_dir),
+        )
+        cloudformation = boto3.resource(
+            'cloudformation',
+            region_name=self._aws_region,
+        )
+        stack_filter = cloudformation.stacks.filter(StackName=self._cluster_id)
+        filtered_stacks = stack_filter.all()
+        [stack] = list(filtered_stacks)
+        stack_id = stack.stack_id
+        launcher = get_launcher(  # type: ignore
+            config=validated_launch_config,
+        )
+        # This matches what happens in
+        # ``dcos_launch.aws.DcosCloudformationLauncher.create``.
+        launcher.config['stack_id'] = stack_id
+        key_helper_details = launcher.key_helper()  #
+        zen_helper_details = launcher.zen_helper()
+        launcher.config['temp_resources'] = {
+            **key_helper_details,
+            **zen_helper_details,
+        }
+        launcher.delete()