+++ CHANGELOG.rst
@@ -6,6 +6,8 @@ Changelog
 Next
 ----
 
+* Add ``minidcos aws destroy`` and ``minidcos aws destroy-list`` commands.
+
 2019.06.10.0
 ------------